+++ azdev/utilities/__init__.py
@@ -22,7 +22,8 @@ from .const import (
     IS_WINDOWS,
     ENV_VAR_TEST_MODULES,
     ENV_VAR_TEST_LIVE,
-    ENV_VAR_VIRTUAL_ENV
+    ENV_VAR_VIRTUAL_ENV,
+    EXT_REPO_NAME
 )
 from .display import (
     display,
@@ -30,7 +31,12 @@ from .display import (
     heading,
     subheading
 )
+from .git_util import (
+    diff_branches,
+    filter_by_git_diff
+)
 from .path import (
+    extract_module_name,
     find_file,
     find_files,
     make_dirs,
@@ -54,6 +60,8 @@ __all__ = [
     'output',
     'heading',
     'subheading',
+    'diff_branches',
+    'filter_by_git_diff',
     'call',
     'cmd',
     'py_cmd',
@@ -67,7 +75,9 @@ __all__ = [
     'ENV_VAR_TEST_MODULES',
     'ENV_VAR_TEST_LIVE',
     'ENV_VAR_VIRTUAL_ENV',
+    'EXT_REPO_NAME',
     'IS_WINDOWS',
+    'extract_module_name',
     'find_file',
     'find_files',
     'make_dirs',