+++ HISTORY.rst
@@ -3,6 +3,19 @@
 Release History
 ===============
 
+0.1.5
+++++++
+* `azdev extension add/remove`: Added ability to supply wildcard (*) to install all available dev extensions
+                                and remove all installed dev extensions.
+* `azdev setup`: Added ability to install all extensions using `--ext/-e *`. Added ability to install CLI edge build
+                 with `--cli/-c EDGE`.
+* `azdev style/test/linter`: Add special names CLI and EXT to allow running on just CLI modules or just extensions.
+                             Add new argument group `--tgt`, `--src`, `--repo` to allow checking only modules or
+                             extensions which have changed based on a git diff.
+* `azdev verify license`: Command will not check any dev-installed CLI and extension repos. Previously, it only
+                          checked the CLI repo.
+* Added new commands `azdev cli/extension generate-docs` to generate sphinx documentation.
+
 0.1.4
 ++++++
 * `azdev linter`: Fix issue with help example rule.